Use 'i64' to represent numbers instead of 'u64'.
- This change is to be able to have negative results without using 'i128'.

+++ src/operator.rs
@@ -88,7 +88,7 @@ impl Operator {
         }
     }
 
-    pub fn eval(&self, operands: Vec<i128>) -> Option<i128> {
+    pub fn eval(&self, operands: Vec<i64>) -> Option<i64> {
         if operands.len() != self.num_operands() as usize {
             return None;
         }
